added jwt and locked the apis

+++ tests/endpoint_crud.py
@@ -1,24 +1,41 @@
+import json
+
 
 class entriescrud():
 
     def __init__(self, app):
         self.test_client = app.test_client()
 
+    def signup_get_token(self):
+        self.test_client.post('/api/v1/auth/signup',
+                              data=dict(phonenumber="122",
+                                        password="112",
+                                        name="kool"
+                                        )
+                              )
+        response = self.test_client.post('/api/v1/auth/login',
+                                         data=dict(phonenumber="122",
+                                                   password="112"
+                                                   )
+                                         )
+        return json.loads(response.get_data(as_text=True))["access_token"]
+
     # Fetch all entries
 
     def test_fetch_all_entries(self):
-        return self.test_client.get('/api/v1/entries')
+
+        return self.test_client.get('/api/v1/entries', headers={'Authorization': 'Bearer ' + self.signup_get_token()})
 
     # Fetch single entry
 
     def test_fetch_one_entries(self):
         self.submit_entry("mm", "mm")
-        return self.test_client.get('/api/v1/entries/0')
+        return self.test_client.get('/api/v1/entries/0', headers={'Authorization': 'Bearer ' + self.signup_get_token()})
 
     # Fetch empty entries
 
     def fetch_one_empty_entries(self):
-        return self.test_client.get('/api/v1/entries/100')
+        return self.test_client.get('/api/v1/entries/100', headers={'Authorization': 'Bearer ' + self.signup_get_token()})
 
     # post put entry with the provided args
 
@@ -26,7 +43,7 @@ class entriescrud():
         return self.test_client.post('/api/v1/entries',
                                      data=dict(entry_title=entry_title,
                                                entry=entry
-                                               )
+                                               ), headers={'Authorization': 'Bearer ' + self.signup_get_token()}
                                      )
 
     # post entry with some form key  missing
@@ -36,12 +53,12 @@ class entriescrud():
             return self.test_client.post('/api/v1/entries',
                                          data=dict(entry="entry",
                                                    entry_date="entry_date"
-                                                   )
+                                                   ), headers={'Authorization': 'Bearer ' + self.signup_get_token()}
                                          )
         if missing_form_name == "entry":
             return self.test_client.post('/api/v1/entries',
                                          data=dict(entry_title="entry_title"
-                                                   )
+                                                   ), headers={'Authorization': 'Bearer ' + self.signup_get_token()}
                                          )
     # put entry with the provided args
 
@@ -50,7 +67,7 @@ class entriescrud():
         return self.test_client.put('/api/v1/entries/0',
                                     data=dict(entry_title=entry_title,
                                               entry=entry
-                                              )
+                                              ), headers={'Authorization': 'Bearer ' + self.signup_get_token()}
                                     )
     # put entry with some form key  missing
 
@@ -59,18 +76,18 @@ class entriescrud():
         if missing_form_key == "entry_title":
             return self.test_client.put('/api/v1/entries/0',
                                         data=dict(entry="entry"
-                                                  )
+                                                  ), headers={'Authorization': 'Bearer ' + self.signup_get_token()}
                                         )
         if missing_form_key == "entry":
             return self.test_client.put('/api/v1/entries/0',
                                         data=dict(entry_title="entry_title"
-                                                  )
+                                                  ), headers={'Authorization': 'Bearer ' + self.signup_get_token()}
                                         )
     # delete entry with the provided args
 
     def submit_delete_entry(self):
         self.submit_entry("mm", "mm")
-        return self.test_client.delete('/api/v1/entries/0')
+        return self.test_client.delete('/api/v1/entries/0', headers={'Authorization': 'Bearer ' + self.signup_get_token()})
 
     def submit_delete_empty_entry(self):
-        return self.test_client.get('/api/v1/entries/200')
+        return self.test_client.get('/api/v1/entries/200', headers={'Authorization': 'Bearer ' + self.signup_get_token()})
